requirements.txt
# Core dependencies
pydantic==2.5.2
notion-client==2.0.0
python-dotenv==1.0.0
aiohttp==3.9.1
loguru==0.7.2
# requests==2.31.0 # Potentially unused or transitive
fastapi==0.104.1
uvicorn==0.23.2
python-dateutil==2.8.2
# jwt==1.3.1 # Potentially unused or transitive
# PyJWT==2.7.0 # Potentially unused or transitive; python-jose is kept as more likely for FastAPI

# Database and validation
sqlalchemy==2.0.23
alembic==1.12.1
email-validator==2.0.0
pydantic-settings==2.0.3

# AI Providers
openai==0.27.8
anthropic==0.2.8

# Vector database and embeddings
pgvector==0.2.4
sentence-transformers==2.2.2
numpy==1.24.3
nltk==3.8.1
colorama==0.4.6

# Integration services
pyairtable==1.4.0
woocommerce==3.0.0
aiofiles==23.1.0
beautifulsoup4==4.12.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Docker deployment
python-multipart==0.0.6
